+++ docker/changes.md
@@ -32,6 +32,8 @@ v0.0.10 - Testing vectorised loads on elementwise kernels.
 
 v0.0.11 - Testing online softmax (new test for casual mask added).
 
+v0.0.12 - Testing shuffle reduction.
+
 ### CUDA Smoke Tests
 
 v0.0.1 - First smoke test for elementwise addition and multiplication.
@@ -77,6 +79,8 @@ v0.0.12 - Different BK on the matmul.
 v0.0.13 - Elementwise vectorised loads. 
 
 v0.0.14 - Online softmax
+
+v0.0.15 - Shuffle reduction.
 
 ### Profiling Layer 1